CODEX

+++ dataBusiness/contacts/src/contacts_type.cpp
@@ -71,7 +71,7 @@ void ContactsType::PrepopulateCommonTypes(std::shared_ptr<OHOS::NativeRdb::RdbSt
     vectorType.push_back(ContentTypeData::CAMCARD);
     vectorType.push_back(ContentTypeData::SIP_ADDRESS);
     unsigned int size = vectorType.size();
-    for (unsigned int i = 0; i < size; i++) {
+    for (int i = 0; i < (int)size; i++) {
         std::string typeValue = vectorType[i];
         int error = LookupTypeId(rdbStore, typeValue);
         if (error == RDB_EXECUTE_FAIL) {
